Get all integers to 32 bits and correct resulting segfault
For consistency, I set all integers to uint32_t (except of course the ones
that should explicitly be on 8 bits). Following that, there was a segfault
caused by the blur functions. The problem was the following:

Declare uint32_t x = 0; uint32_t r = 10;
If you do x-r, the result is -10 BUT the result of (x-r) >= 0 is 1!
I understand that unsigned - unsigned is still unsigned, so the test fails.
It's just weird that the pure x-r operation (no test) gives -10. There's
probably implicit casts and priorities that I don't fully understand right now.
Anyway, I solved that by casting to signed integer before the test, which works
but might cause problems with other compilers.