Fix best scale calculation
The problem with the previous implementation was that the patterns stabilized
quickly. This came from a wrong way to compute the best scale to update the
image. The program used to do that:
- Compute n variation arrays, n being the number of scales. Each variation array
has size h*w (i.e. like the image).
- Find the smallest variation for each variation array. This gives n values.
- Find the smallest value among them. This gives one scale.
- Use this scale to update all pixels of the image.
This is wrong! Or at least this is not what the algorithm is supposed to do.
At one given timestep, all pixels are updated on the same scale. What the
program should do is:
- Compute the n variation arrays.
- For each pixel of the image, find which variation *at that point only* is the
smallest.
- Use the scale that corresponds to this variation to update the pixel.
So now, each pixel is updated on a specific scale.