Migrate to SDL and C++
Changing the display library from gifsave to SDL required a migration to C++.
It's not really C++ though, it's just the old C code with a few corrections
to make it compilable by g++. The use of SDL is also very basic, since it's the
first time I work with it.

During the adaptations to C++, I made a couple of other changes, including:
- All image arrays are now one-dimensional, it's much simpler to process them
  like that most of the time.
- The problem with displaying not-square images is gone, probably thanks to the
  one-dimensional arrays and proper indexing. I still don't know exactly where
  the problem was.
- Generated pixel values are now in the interval [0; 1] instead of [-1; 1].
  It's simpler this way and it doesn't change much to the functions.
- Three steps of the main algorithm (compute variations, find best scale,
  update image) are now in the same function instead of three separate ones.
  This reduces the need for intermediate arrays, the memory usage, and the risk
  of segmentation faults. Of course it makes the program a bit harder to
  understand since all the steps are kind of mashed together now.

There are still segmentation faults when the image is too big, so it would be
good to optimize the number of arrays needed.